tech-info/requirements.txt
Flask
requests
python-Wappalyzer
beautifulsoup4
